Move code to appropriate packages based PR feedback
