+++ internal/nonkube/bundle/install.sh.template
