+++ internal/nonkube/bundle/self_extract.sh.template
