+++ internal/utils/command_test.go
@@ -1,4 +1,4 @@
-package nonkube
+package utils
 
 import (
 	"testing"
