+++ internal/utils/tarball.go
@@ -1,4 +1,4 @@
-package nonkube
+package utils
 
 import (
 	"archive/tar"
